first cut at parser for the list of ministerial committees
Parses and creates internal data structure just fine, but
the output format is not done. Will happen later today.

URL is included in the script as it seems to be a document
posted as part of a news item, rather than in a static
place, so finding it will be a manual job.