Convert non-ASCII characters in speaker names to HTML entities
In the latest Scottish Parliament data there are speaker names that
contain curly apostrophes, em dashes and other non-ASCII characters.
This commit makes sure that they are converted to HTML entities in the
output.  (This isn't required for the text of speeches since they are
emitted with Twig's sprint, which uses the output filter.)