alert people by email if they are signed up for an existing alert
If someone not logged in, or someone with a different email address
attempts to sign up for an existing alert do not display the already
signed up error on the site but instead send an email to the relevant
email address.

Stops people working out if a user is signed up to an alert by
attempting to sign them up.