basic template structure for new MP page
Moving to having the header, footer and page content 'templated' which
will then be included by the file that displays them. All the data will
be set up outside the templates which should only do presentation.

This only includes the templates and not the logic page

also initial rough sass for the layout