make section nav fixed on scrolling
uses compasses magellan component. This means we need to add a cheeky
empty div to the column to keep the main content in place when the nav
is fixed. We also need to set the width of the nav to two columns when
it's fixed otherwise it ends up a bit narrow