add details link to policy votes
turn key_votes into an array of arrays with the array containing the
policy_id as well so we can use it for generating the link.
include member_id as well as person_id as this is needed for PW link.
use these to create a details link
change JS hover code for new class names
Add border and button styling