Switch to using Foundation v4.2.3 via a submodule
Previously we made the assumption that someone had installed the
zurb-foundation gem somewhere accessible, e.g. via
"sudo gem install zurb-foundation" or in an rvm gemset.  However,
there's no need for this assumption, since we can just point to the
version of Foundation that we require as git submodule.

This uses Foundation v4.2.3 which was the version being used
previously on our servers.