Prevent loading in older data rolling back change.
With the new written answers, it's possible for an answer to come in one day,
then a new answer (replacing a holding answer, or a correction) to come on a
later day. This works fine and the answer is replaced, unless you then loaded
in the old date manually, when the old out-of-date answer would be reimported.

Therefore, do not import any data which would move the hdate back in time -
this only applies to new-style written answers, where an ID may occur on more
than one date.