retain search term over multiple redirects
if the link from the search page, or more realistically an email goes
through more than one level of redirect then the search term is lost
from it. Check if it's present and add it back in to the redirect.

Note that we will always lose and anchor part of the URL in this process
as that's all dealt with purely client side and never supplied to the
server.

Fixes #754