Use person ID rather than member ID in database.
Store the person ID in the hansard and pbc_members tables, rather than
the member ID. This makes it easier to update the database with issues
of out-of-date membership changes.

One special case for the election of the NI Speaker which happens during
the course of a day; I can't obviously see an easy solution with the
existing model and it is so rare.