Show weak votes by default if there are no strong votes
It's possible than an MP may only have voted on weak votes associated
with a policy in which case it makes no sense to hide them and force the
user to have to click a button to see any content on the page. In this
situation hide the button and display all the votes.