Update text on the api key request page
This updates the text on this page and adds a new option to the form
for estimated usage. If the user selects a commercial use or says
their usage is likely to be over 50,000 calls a year, it gives them a
message that says they're likely to need to pay.

Requires a migration running on the database to add the new field to
the api_key table;