do not display divisions that a policy abstains on
On PublicWhip it's possible to list a division as abstained to indicate
that it's related to a policy but that it doesn't count towards an MPs
position on the policy. This is probably just a bit to confusing for
TheyWorkForYou so don't include them in the list of divisions for a
policy.

Fixes #830