Float .debate-speech__links instead of abs position
Part of #745 and also a belated fix for #678.

By floating the speaker name and speech left, and the meta links right,
we don’t have to worry about absolutely-positioned links overlapping
the bottom of the speech.

Also took the opportunity to simplify some of the breakpoint behaviour
and use percentages for widths in the squidgy 800-1100px screen range.