ignore case differences when loading constituencies
Because mysql's default collation is case insensitive the replace into
is triggered by case differences in names becuase the index sees
Berwick-upon-Tweed and Berwick-Upon-Tweed as the same. This means the
for the second one we update main_name to 0 which means no results are
found by the constituency name search.

Fixes #856