swap search term and glossary highlighting order
Because of the way the search engine highlighting works it will add
highlighting to text that is inside tags, e.g. in the title tag of a
glossary link. The glossary highlighting doesn't do this so swap the
order we do these in.

This also tweaks the search highlight matching RegExp so that phrases do
not match inside tags, e.g. if the search phrase is in a title attrib

Fixes #912