Fix uninitialized value when a new policy divisions is added
Whe a new policy division was added there was no current version of the
division to compare it agains to see if the vote had changed from strong
to weak so an uninitialized value warning was generated.
We now check to see if there is an existing division before doing to
comparison.

Also add some comments to say what all this is doing