Always process checks for strong votes in motions
If we skip the whole of the vote processing of a motion then it's
possible that a policy that only contains motions that are in other
policies never has the has_strong_vote mp info set and so always
displays as not having enough information to decide a position.

This moves the check into the motions loop and only skips the setting of
the vote information in the persondivisionvotes table. We also move
setting the motion_seen code to after the vote processing to avoid
always skipping updates.

Fixes #961