Avoid advanced search options flash on page load
Fixes #1001.

The advanced search options are now shown and hidden based on
the presence of a `.js` class on the `<html>` element, which
gets set by Modernizr very early in pageload. This should stop
the element appearing momentarily when the search page loads.

We also split the form up into two parts - a short one is displayed
above the results, and a longer one (with the advanced options)
displayed after the results. If javascript is available, the short
form is replaced with the longer one and the advanced options are hidden,
to be later revealed naturally, using the toggle button.