Promote MP responsiveness on MP profile page
Loads all the responsiveness strings into a $wtt_strings variable that's
available to the template and then uses a year variable in the template
to show/hide the promotional banner.

Also extracts the list of WTT stats years into an array at the top of
mp/index.php to make it a bit easier to add new years to this.

And adds 2015 while we're at it which means all this will just work for
2015 once the stats are available.