reduce queries to generate policy vote summaries
fetch the details for all policies in two queries rather than two per
policy and then pass that in to the policies object. And then split the
summary generation out to its own method so it can be called with the
data passed in to policies.

means only 2 queries rather that 2 * $number of policies.