store vote totals and result for divisions
Because we are going to be displaying summary information about votes
next to them calculate the summary details (votes for/against etc) as we
import the vote data rather than have to do it on the fly. At also
avoids having to right some fairly awful SQL in order to calculate
things on the fly.

As vote data only comes in via one script there should be no
problem with consistency.