Remove mySociety header
We think people are clicking the mySociety logo in the header
in an attempt to contact their MP (and thereby causing lots of
erroneous emails to user support via the mysociety.org contact
form). Removing the logo from the header should make the *real*
path for contacting your MP more obvious.

Fixes #1246.