allow users to login/register using facebook
Add the option to use facebook login instead of email login. It'll use
an existing email account if it exists otherwise it created a new user
account. We don't send confirmation emails for FB login as we assume
that the FB email is correct.

Facebook logins use a different cookie to distinguish them from email
logins and hence enable us to do the correct checks.