hide annotations section if there are no annotations
Becuase we're not allowing new annotations there's no point drawing
attention to the annotations section on pages that don't already have an
annotation so remove the 'No annotations' message. This doesn't remove
the div so there's a stray row on individual entry pages at the moment.
This is becuase the way the code is structured doesn't make it possible
to determine in advance if there are any comments and this is just a
quick fix.

To make this work we'd need to extract the method that get the comments
outside the code that displays them and then probably pass that directly
to the template rather than the comment object.