add missing pg value to user edit page
Otherwise we don't know we are editing the page and the update fails.

Return some more information from `getUpdateDetails` that is
required to display the user information page after a succesful update.

Just pass in the user object to getUpdateDetails rather than rely on a
global or passing in individual values.

Moves facebook_user function into USER class instead of THEUSER class as
it's used in USER so we need to make sure it exists.

Add some basic tests for the /user pages