Standardise Facebook sharing for MP votes and speeches
We were using the Facebook Feed Dialog (`FB.ui({ method: 'feed' })`)
via `.fb-share-button-tracked`, to share images of votes, but most of
the Feed Dialog’s parameters (including `picture`) were deprecated by
Facebook in July 2017.

As it happens, thanks to the `/policy_set_png` og:image tag in the
shared `/votes` URL, Facebook was pulling the right image out anyway,
so this commit just pulls out all that `.fb-share-button-tracked` stuff
and replaces it with the `.js-facebook-share` behaviour we introduced
for sharing speeches in debates (which uses the Facebook Share Dialog,
rather than the Facebook Feed Dialog).