Refreshed header for MP/person profile pages
* Displays social media links more prominently, fixing #1133.
* Only displays the current, or most recently ended, membership.
* Displays the position an MP might be "known for", ie: non-committee
  positions such as ministerial roles.
* Removes the house-based background photo.
* Moves profile photo attribution into footer.

The header code assumes images in `mpsL/` are 120px or wider, 3:4
portrait aspect ratio, which isn’t currently true on the live version
of theyworkforyou.com. To solve this, we’re going to move the current
“large” (but not large enough) images from `mpsL/` into `mps`, so
they’ll be displayed at their natural size, with the “blurred”
background wrapper, and then we can upload *properly* large images to
`mpsL/` at some point in the near future.

We also took this opportunity to simplify getMostRecentMembership().