Remove reference to a non-existent button
A user wrote in to say "I have followed the instructions however there is not a button named “I want to write to this Lord” as described." and was right, there isn't. This commit changes the instructions to refer to what's actually on the page (which changed, I presume, in a long-ago redesign).