Refactor how type of representative is detected.
Instead of setting variables such as `peer` and `mp` to 1, this now sets
`representative_type` to `peer`, `mp` etc.

This then means chained if/else statements can be replaced with a switch.

Also moves initialisation of `$people` in member lists away from being
done dynamically based on a variable value, and just does it in the
switch statement. The end result is the same.