Fix multiple ORs in email alerts.
Xapian has changed how it brackets its terms; this code assumed it would
be `(((B123 OR B124) OR B125) OR B126)` whereas now it is `(B123 OR B124
OR B125 OR B126)`. As the old regex still matched on the `B\d+` part, an
email alert would be left with lots of "OR"s in the alert titles.