Fix date format for NI Speaker exceptions in the Xapian indexer
Although the database stores dates as YYYY-MM-DD, the hypens are removed
between the query being run and the data passed into `get_person`, so
these conditions weren't matching.

Noting that as per comments in the code, the same checks are made in both
index.pl and hansardlist.php but the code in the latter uses the
hyphenated dates and did not need changing.