Fix conflicting no-js/js html/body classes
Looks like the `no-js` class on the html element is a leftover from
Modernizr.js, which was removed in 2019 (d12d92b5b).

We have styles which rely on a `js` class being added, so I’ve kept
that, but moved it to the root HTML element, and moved the piece of
JavaScript which adds the `js` class to inline in the header, so it
runs earlier than its previous location in main.js.