Fix conflicting no-js/js html/body classes
Looks like the `no-js` class on the html element was a leftover from
Modernizr.js, which was removed in 2019 (d12d92b5bdeb36f8391820a5702971b44cbb36b0).

We have styles which rely on a `js` class being added to the body
element, so I’ve kept that, but moved the piece of JavaScript which
adds the `js` class to right after the opening of the body element,
which should make it run earlier than its previous location in main.js.

Part of #1421.