Scrolling MP page sidebars with CSS position:sticky
These sidebars were the last place we were using Foundation.js
(for its Magellan sticky positioning module) so replacing them with
position:sticky means we can drop all Foundation JavaScript.

Position:sticky requires a full height parent element, so we upgrade
the MP page .person-panels to use flex layout, which gives us full
height columns automatically.

One thing we’ve lost with the removal of Foundation Magellan is
highlighting of the "current" sidebar link. We could do this via the
IntersectionObserver API if we wanted, but for now I’m not sure it’s
worth it.

Part of #1421.