[Docker] First pass at a Docker development environment
This builds an image with the required packages based on Debian Buster
and deliberately sets the entrypoint in the Compose file rather than
the Dockerfile so that we can simply map in our working copy and start
everything up.

The `docker-entrypoint.sh` script takes care of setting up the config
file, installing the composer and bundler dependencies and starting
Apache in the foreground.

This will mean that the vendored files will persist locally; these do
not use Docker volumes but rather the local filesystem; I have used a
Docker volume for the Xapian index.