[Docker] Restructure package installation in Dockerfile
The first iteration of this Dockerfile combined the contents of the
`conf/packages` file with some packages needed within the container to
run the application (Apache, etc) plus some build dependencies.

This is a bit of a maintenance pain, so I've instead separated this into
three sections, one that installs service packages needed, one for build
dependencies not currently specified in `conf/packages` and a final step
that copies in `conf/packages` and iterates over that.

This approach adds layers and makes the build slightly slower in favour
of slightly improved maintainability - we only need to make edits to
`conf/packages` to add runtime dependencies, and it's immediately clear
what is needed for build and what for the service.