Party comparisons moved from all-time to cohorts.
It is not always appropriate to compare MPs to an all-time score.
This commit introduces new tables and classes have created to compare
MPs only to those who were members at the same time as them.
This also includes explicit handling for if a party-switcher is
judged by their original or final party.