Add Support for Redis Sentinel
This adds support for Redis Sentinel mediated connections. Set the
constant `REDIS_SENTINELS` to an array of IP addresses for the sentinel
servers. `REDIS_SERVICE_NAME` will also need to be set to the name of
the service. `REDIS_SENTINEL_PORT` has also been provided, although we
assume that each sentinel is on the same port.

If `REDIS_SENTINELS` is not set, then we'll fall back to making a direct
connection using the usual variables.