Extend GitHub Actions with Redis and Sentinel
This swaps the zhulik/redis-action container with redis and
bitnami/redis-sentinel so that we can test either direct connections to
redis, or indirect via sentinel.

The tests can be configured to use either and will currently follow the
`general-example` file, but it would be good to test both types of
connection regardless of which was used to run the complete set of unit
tests subsequently.