Update Docker Compose environment
This adds support for Redis/Sentinel and updates the version of MariaDB
in use to match that used in Debian Bullseye.

This also removes the local Xapian volume mounted at `searchdb/`. In
cases where this directory doesn't contain a Xapian index - which will
be in all cases on a fresh clone - tests will fail even when the xapian
group is excluded.

It also adds a local volume for `vendor/` to avoid polluting the local
filesystem and potentially causing confusion.

Probably the best way to use this environment to run the tests is to
create it from scratch each time then remove everything afterwards. That
way you'll be confident of a clean run.