enable using popovers for glossary instead of links to page
if we're on a section page then as well as glossarising the content of
the page, return a list of the terms with their content that we can then
inject as hidden elements into the page. Enables turning the glossary
entry into a popup rather than linking to a fresh page.